ai-chatbot-clone/requirements.txt
flask
flask-cors
bcrypt
mysql-connector-python
requests